Scroll with apartments and their data in search section
