+++ app/src/main/java/ch/epfl/sweng/hostme/Apartment.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class Apartment {
 
     private String address;
-    private String area;
+    private int area;
     private boolean available;
     private Date currentLease;
     private String bath;
@@ -25,7 +25,7 @@ public class Apartment {
 
     public Apartment() {}
 
-    public Apartment(String address, String area, boolean available, Date currentLease, String bath, int deposit,
+    public Apartment(String address, int area, boolean available, Date currentLease, String bath, int deposit,
                      boolean furnished, String image_path, String kitchen, String laundry, String lid,
                      String name, int occupants, boolean pets, String proprietor, int rent, String room,
                      int utilities) {
@@ -185,11 +185,11 @@ public class Apartment {
         this.utilities = utilities;
     }
 
-    public String getArea() {
+    public int getArea() {
         return area;
     }
 
-    public void setArea(String area) {
+    public void setArea(int area) {
         this.area = area;
     }
 }
