+++ app/src/main/java/ch/epfl/sweng/hostme/MenuActivity.java
@@ -27,6 +27,7 @@ public class MenuActivity extends AppCompatActivity {
         binding = ActivityMenu1Binding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
+
         BottomNavigationView navView = findViewById(R.id.nav_view);
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
