+++ app/src/main/res/layout/list_item.xml
@@ -1,10 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/relativeLayout"
     android:layout_width="match_parent"
     android:layout_height="?android:attr/listPreferredItemHeightLarge"
     android:background="@drawable/border">
+
+    <ImageView
+        android:id="@+id/apartment_image"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
 
     <TextView
         android:id="@+id/list_price"
@@ -40,4 +45,4 @@
         tools:layout_editor_absoluteX="300dp"
         tools:layout_editor_absoluteY="16dp" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</RelativeLayout>
