+++ app/src/main/res/layout/login_page.xml
@@ -64,7 +64,7 @@
         app:layout_constraintTop_toBottomOf="@+id/userName"
         tools:ignore="LabelFor" />
 
-    <Button
+    <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/logInButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -78,7 +78,7 @@
         app:layout_constraintTop_toBottomOf="@+id/pwd"
         tools:ignore="TextContrastCheck" />
 
-    <Button
+    <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/signUpButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -92,7 +92,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/logInButton" />
 
-    <Button
+    <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/forgotPassword"
         android:layout_width="207dp"
         android:layout_height="54dp"
