+++ app/src/main/res/layout/recycler_view.xml
@@ -1,9 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.recyclerview.widget.RecyclerView
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
+
+<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
     android:scrollbars="vertical"
-    android:id="@+id/recyclerView">
-
+    android:id="@+id/recyclerView"
+    tools:context=".ui.search.SearchFragment">
 </androidx.recyclerview.widget.RecyclerView>
+
+</LinearLayout>
